+++ src/main/java/org/folio/inventory/resources/Items.java
@@ -6,7 +6,6 @@ import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.BodyHandler;
-import org.folio.inventory.storage.external.CollectionResourceClient;
 import org.folio.inventory.common.WebContext;
 import org.folio.inventory.common.api.request.PagingParameters;
 import org.folio.inventory.common.domain.MultipleRecords;
@@ -14,6 +13,8 @@ import org.folio.inventory.common.domain.Success;
 import org.folio.inventory.domain.Item;
 import org.folio.inventory.domain.ItemCollection;
 import org.folio.inventory.storage.Storage;
+import org.folio.inventory.storage.external.CollectionResourceClient;
+import org.folio.inventory.support.JsonArrayHelper;
 import org.folio.inventory.support.http.client.OkapiHttpClient;
 import org.folio.inventory.support.http.client.Response;
 import org.folio.inventory.support.http.server.*;
@@ -23,7 +24,6 @@ import java.lang.invoke.MethodHandles;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -32,7 +32,6 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static org.folio.inventory.common.FutureAssistance.allOf;
-import org.folio.inventory.support.JsonArrayHelper;
 
 public class Items {
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
@@ -247,7 +246,7 @@ public class Items {
 
   private Item requestToItem(JsonObject itemRequest) {
     List<String> pieceIdentifiers;
-    pieceIdentifiers = JsonArrayHelper.toListOfStrings(itemRequest.getJsonArray("pieceIdentifiers"));    
+    pieceIdentifiers = JsonArrayHelper.toListOfStrings(itemRequest.getJsonArray("pieceIdentifiers"));
     String status = getNestedProperty(itemRequest, "status", "name");
     List<String> notes;
     notes = JsonArrayHelper.toListOfStrings(itemRequest.getJsonArray("notes"));
@@ -266,6 +265,7 @@ public class Items {
       pieceIdentifiers,
       itemRequest.getString("numberOfPieces"),
       itemRequest.getString("instanceId"),
+      itemRequest.getString("holdingsRecordId"),
       notes,
       status,
       materialTypeId,